requirements.txt
# Core ML and Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# PyTorch (for speech processing and diarization)
torch>=1.11.0
torchaudio>=0.11.0

# Speech Processing and NLP
soundfile>=0.12.0
pyannote.audio>=2.1.0

# Transformers and NLP
transformers>=4.20.0
transformers-interpret>=0.9.0

# Additional dependencies that may be required by pyannote.audio
speechbrain>=0.5.12

# Optional: For better performance with audio processing
librosa>=0.9.0

# Environment and Progress Tracking
python-dotenv>=0.19.0
tqdm>=4.64.0
pydub
faster-whisper
groq

# For embedding visualization
umap-learn>=0.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
scipy>=1.7.0